requirements.txt
# In requirements.txt
discord.py==2.3.2
python-dotenv
discord.py>=2.3.2
pymongo>=4.5.0
dnspython>=2.2.1



